requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==24.2.0
contourpy==1.3.1
cycler==0.12.1
einops==0.8.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.9
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pillow==11.0.0
propcache==0.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
PyYAML==6.0.2
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
yarl==1.18.0
